refactor: rename Toggle to ToggleRefinement
