+++ packages/react-instantsearch/connectors.js
@@ -45,7 +45,9 @@ export {
 } from './src/connectors/connectSearchBox.js';
 export { default as connectSortBy } from './src/connectors/connectSortBy.js';
 export { default as connectStats } from './src/connectors/connectStats.js';
-export { default as connectToggle } from './src/connectors/connectToggle.js';
+export {
+  default as connectToggleRefinement,
+} from './src/connectors/connectToggleRefinement.js';
 export {
   default as connectStateResults,
 } from './src/connectors/connectStateResults.js';
