+++ packages/react-instantsearch/src/connectors/connectToggleRefinement.js
@@ -41,9 +41,9 @@ function cleanUp(props, searchState, context) {
 }
 
 /**
- * connectToggle connector provides the logic to build a widget that will
+ * connectToggleRefinement connector provides the logic to build a widget that will
  *  provides an on/off filtering feature based on an attribute value. Note that if you provide an “off” option, it will be refined at initialization.
- * @name connectToggle
+ * @name connectToggleRefinement
  * @kind connector
  * @requirements To use this widget, you'll need an attribute to toggle on.
  *
